Regex chg to match new Logcat tag
Logcat tag change in the CalcyIV logcat output in v.3.23a build 32355. 
Have to chg the regex to match the j .  Nothing else change in the regex, as far as I can see (since the big change in v.3.21 build 32150)
Sometime in the beginning of March 2020.
Logcat tag change from i to j.