The rename bug in PoGo v171
We will use the touch-paste parameter as a workaround for the rename to function.
It will simulate a finger tab on the delete key for some milliseconds. My slow Samsung J6 need at least 1000.
Then send the key command PASTE.
Then a finger tab on the DONE.
Coordinates: 
As keenefisimatenten says, use the edit_box to point to the DELETE/BACKSPACE key coordinates. 
Add another coordinate named "done" next to edit_box or rename_ok. Let it point to the "DONE" key, you normally press when finish paste some text.
Then rename_ok should still point to the green OK button for the new Pokemon name ;-)
Remember to use the the parameter --touch-paste