Delete needs TapAndHold
!!!REMEMBER to update/check you location/coordinate for the key named "edit_box". It should map/press the DEL key on you rename keyboard. Probably there is some text behind it like "only used with --touch-paste"
After PoGo v0.173
If you are using standard keyboard.
When rename, to delete the current name, you need to press and hold the DEL key until all is deleted. (or press multi times)
I thing before v0.171 (the rename bug) the text was marked when you enter the rename box, so you only need to press the DEL key once.